requirements.txt
# Core Flask
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.4
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1

# Database
SQLAlchemy==2.0.36
psycopg2-binary>=2.9.9,<3.0.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Security
cryptography>=41.0.0
itsdangerous==2.2.0
blinker==1.8.2
Flask-Login==0.6.3

# Forms
WTForms==3.1.2
email-validator==2.1.0.post1

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2024.1

# API and HTTP
requests==2.31.0
Flask-Cors==4.0.0